+++ main.tf
@@ -146,7 +146,7 @@ module "core_config" {
 
   dns_resource_group_lookup = var.dns_resource_group_lookup
 
-  logging = { control_plane = { log_analytics = merge(var.logging.control_plane.log_analytics, { workspace_id = module.cluster.control_plane_log_analytics_workspace_id }), storage_account = var.logging.control_plane.storage_account }, workloads = var.logging.workloads }
+  logging = merge(var.logging, { control_plane = { log_analytics = merge(var.logging.control_plane.log_analytics, { workspace_id = module.cluster.control_plane_log_analytics_workspace_id }), storage_account = var.logging.control_plane.storage_account } })
   storage = local.storage
 
   core_services_config = local.core_services_config